Partially document use of LLM
3 of the last prompts used.
Models unknown but maybe Claude, or GPT.
Most of the major initial prompts are not documented.

Prompt N 1:
On mobile, I would like to allow the user to drag up and down on the 3d cyclinders and have them rotate

Prompt N + 2:
Instead of applying the rotation to the cylinder, can you hook that into changing the value of the slider please.

Prompt N + 3:
On a mobile, when the screen is too narrow, I would like the view to be zoomed out by 30% so that the content can fit on the screen.  Could you modify the code please.  Thank you